@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -31,19 +32,23 @@ type DataPolicyManifestsClient struct {
 // NewDataPolicyManifestsClient creates a new instance of DataPolicyManifestsClient with the specified values.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewDataPolicyManifestsClient(credential azcore.TokenCredential, options *arm.ClientOptions) *DataPolicyManifestsClient {
+func NewDataPolicyManifestsClient(credential azcore.TokenCredential, options *arm.ClientOptions) (*DataPolicyManifestsClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := options.Endpoint
-	if len(ep) == 0 {
-		ep = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &DataPolicyManifestsClient{
-		host: string(ep),
-		pl:   armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options),
+		host: ep,
+		pl:   pl,
 	}
-	return client
+	return client, nil
 }
 
 // GetByPolicyMode - This operation retrieves the data policy manifest with the given policy mode.
@@ -86,7 +91,7 @@ func (client *DataPolicyManifestsClient) getByPolicyModeCreateRequest(ctx contex
 
 // getByPolicyModeHandleResponse handles the GetByPolicyMode response.
 func (client *DataPolicyManifestsClient) getByPolicyModeHandleResponse(resp *http.Response) (DataPolicyManifestsClientGetByPolicyModeResponse, error) {
-	result := DataPolicyManifestsClientGetByPolicyModeResponse{RawResponse: resp}
+	result := DataPolicyManifestsClientGetByPolicyModeResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.DataPolicyManifest); err != nil {
 		return DataPolicyManifestsClientGetByPolicyModeResponse{}, err
 	}
@@ -101,16 +106,32 @@ func (client *DataPolicyManifestsClient) getByPolicyModeHandleResponse(resp *htt
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - DataPolicyManifestsClientListOptions contains the optional parameters for the DataPolicyManifestsClient.List
 // method.
-func (client *DataPolicyManifestsClient) List(options *DataPolicyManifestsClientListOptions) *DataPolicyManifestsClientListPager {
-	return &DataPolicyManifestsClientListPager{
-		client: client,
-		requester: func(ctx context.Context) (*policy.Request, error) {
-			return client.listCreateRequest(ctx, options)
+func (client *DataPolicyManifestsClient) List(options *DataPolicyManifestsClientListOptions) *runtime.Pager[DataPolicyManifestsClientListResponse] {
+	return runtime.NewPager(runtime.PageProcessor[DataPolicyManifestsClientListResponse]{
+		More: func(page DataPolicyManifestsClientListResponse) bool {
+			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		advancer: func(ctx context.Context, resp DataPolicyManifestsClientListResponse) (*policy.Request, error) {
-			return runtime.NewRequest(ctx, http.MethodGet, *resp.DataPolicyManifestListResult.NextLink)
+		Fetcher: func(ctx context.Context, page *DataPolicyManifestsClientListResponse) (DataPolicyManifestsClientListResponse, error) {
+			var req *policy.Request
+			var err error
+			if page == nil {
+				req, err = client.listCreateRequest(ctx, options)
+			} else {
+				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
+			}
+			if err != nil {
+				return DataPolicyManifestsClientListResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return DataPolicyManifestsClientListResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return DataPolicyManifestsClientListResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listHandleResponse(resp)
 		},
-	}
+	})
 }
 
 // listCreateRequest creates the List request.
@@ -134,7 +155,7 @@ func (client *DataPolicyManifestsClient) listCreateRequest(ctx context.Context, 
 
 // listHandleResponse handles the List response.
 func (client *DataPolicyManifestsClient) listHandleResponse(resp *http.Response) (DataPolicyManifestsClientListResponse, error) {
-	result := DataPolicyManifestsClientListResponse{RawResponse: resp}
+	result := DataPolicyManifestsClientListResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.DataPolicyManifestListResult); err != nil {
 		return DataPolicyManifestsClientListResponse{}, err
 	}
